# api/openapi.yaml
openapi: 3.0.0
info:
  title: Kart Challenge API
  description: API for managing products, shopping carts, orders, and coupons.
  version: 1.0.0
servers:
  - url: http://localhost:8080/
paths:
  /:
    get:
      summary: Hello World
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /health:
    get:
      summary: Health Check
      responses:
        '200':
          description: Service Healthy
        '503':
          description: Service Unavailable (e.g., database connection failed)
  /products:
    get:
      summary: Get All Products
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /products/{id}:
    get:
      summary: Get Product by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the product to retrieve
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
  /cart:
    get:
      summary: Get or Create User Cart
      parameters:
        - in: header
          name: DeviceID
          schema:
            type: string
          required: true
          description: Unique identifier for the user's device
      responses:
        '200':
          description: User's cart details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid DeviceID
    delete:
      summary: Clear User Cart
      parameters:
        - in: header
          name: DeviceID
          schema:
            type: string
          required: true
          description: Unique identifier for the user's device
      responses:
        '200':
          description: Cart cleared successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid DeviceID
  /cart/items:
    post:
      summary: Add or Update Item in Cart
      parameters:
        - in: header
          name: DeviceID
          schema:
            type: string
          required: true
          description: Unique identifier for the user's device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - product_id
                - quantity
              properties:
                product_id:
                  type: integer
                quantity:
                  type: integer
      responses:
        '200':
          description: Item added/updated in cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid request or DeviceID
        '404':
          description: Product not found
        '409':
          description: Not enough stock
  /cart/items/{product_id}:
    put:
      summary: Update Quantity of Item in Cart
      parameters:
        - in: header
          name: DeviceID
          schema:
            type: string
          required: true
          description: Unique identifier for the user's device
        - in: path
          name: product_id
          schema:
            type: integer
          required: true
          description: Numeric ID of the product in the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - quantity
              properties:
                quantity:
                  type: integer
      responses:
        '200':
          description: Item quantity updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid request, DeviceID, or quantity
        '404':
          description: Product or Cart not found
        '409':
          description: Not enough stock
    delete:
      summary: Remove Item from Cart
      parameters:
        - in: header
          name: DeviceID
          schema:
            type: string
          required: true
          description: Unique identifier for the user's device
        - in: path
          name: product_id
          schema:
            type: integer
          required: true
          description: Numeric ID of the product to remove
      responses:
        '200':
          description: Item removed from cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '400':
          description: Invalid DeviceID
        '404':
          description: Product or Cart not found
  /orders:
    post:
      summary: Place an Order
      parameters:
        - in: header
          name: DeviceID
          schema:
            type: string
          required: true
          description: Unique identifier for the user's device
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                coupon_code:
                  type: string
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request or empty cart
        '409':
          description: Out of stock items
    get:
      summary: Get All Orders for Device
      parameters:
        - in: header
          name: DeviceID
          schema:
            type: string
          required: true
          description: Unique identifier for the user's device
      responses:
        '200':
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Invalid DeviceID
  /orders/{order_id}:
    get:
      summary: Get Order by ID
      parameters:
        - in: header
          name: DeviceID
          schema:
            type: string
          required: true
          description: Unique identifier for the user's device
        - in: path
          name: order_id
          schema:
            type: integer
          required: true
          description: Numeric ID of the order to retrieve
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid DeviceID or Order ID
        '404':
          description: Order not found or does not belong to device
  /coupons/validate:
    post:
      summary: Validate Coupon Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - coupon_code
              properties:
                coupon_code:
                  type: string
                device_id:
                  type: string
      responses:
        '200':
          description: Coupon validated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  coupon:
                    $ref: '#/components/schemas/Coupon'
                  status:
                    type: string
                  discount_amount:
                    type: number
                    format: float
                  discount_type:
                    type: string
                  message:
                    type: string
        '400':
          description: Invalid request or couponLib code
        '404':
          description: Coupon not found or not valid
components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        image_url:
          type: string
        stock:
          type: integer
          format: int32
    Cart:
      type: object
      properties:
        id:
          type: integer
          format: int64
        device_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
    CartItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cart_id:
          type: integer
          format: int64
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        price_at_addition:
          type: number
          format: float
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        device_id:
          type: string
        status:
          type: string
        total_amount:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        order_id:
          type: integer
          format: int64
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        unit_price:
          type: number
          format: float
    Coupon:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        type:
          type: string
        value:
          type: number
          format: float
        description:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time